+++ nodes/common.js
@@ -22,6 +22,10 @@ module.exports = {
     }
 
     throw new Error("missing 'controller' parameter")
+  },
+
+  clamp: function (v, min, max) {
+    return Math.min(Math.max(v, min), max)
   }
 }