+++ nodes/encoder.js
@@ -164,10 +164,11 @@ module.exports = {
     * @param {number} deviceId  Controller serial number
     * @param {string} address   IPv4 listener address for controller events
     * @param {number} port      IPv4 listener port for controller events
+    * @param {number} interval  Auto-send interval (seconds) - 0 for no auto-send.
     *
     * @return {buffer} 64 byte NodeJS buffer with encoded set-address request
     */
-  SetListener: function (deviceId, { address, port } = {}) {
+  SetListener: function (deviceId, { address, port, interval } = {}) {
     const request = Buffer.alloc(64)
 
     request.writeUInt8(0x17, 0)
@@ -177,6 +178,10 @@ module.exports = {
     ipx.toBuffer(address, request, 8)
 
     request.writeUInt16LE(port, 12)
+
+    if (!Number.isNaN(interval) && interval >= 0 && interval < 256) {
+      request.writeUInt8(interval, 14)
+    }
 
     return request
   },