+++ nodes/uhppoted-set-listener.js
@@ -17,6 +17,7 @@ module.exports = function (RED) {
       const controller = common.resolve(msg.payload)
       const address = msg.payload.address
       const port = msg.payload.port
+      const interval = Number.isNaN(msg.payload.interval) ? 0 : common.clamp(msg.payload.interval, 0, 255)
 
       const emit = function (object) {
         if (!object.updated) {
@@ -26,7 +27,8 @@ module.exports = function (RED) {
         common.emit(node, t, {
           controller: controller.id,
           address,
-          port
+          port,
+          interval
         })
       }
 
@@ -41,7 +43,7 @@ module.exports = function (RED) {
           logger: (m) => { node.log(m) }
         }
 
-        uhppoted.set(context, controller.id, opcodes.SetListener, { address, port }, controller.address, controller.protocol)
+        uhppoted.set(context, controller.id, opcodes.SetListener, { address, port, interval }, controller.address, controller.protocol)
           .then(object => { emit(object) })
           .then(done())
           .catch(err => { error(err) })